+++ software-pya1/cmodules/vt100.h
@@ -52,7 +52,7 @@ void lcd_reset();
 void driver_init();
 void driver_setup();
 
-void driver_print(const unsigned char* text, const uint32_t len, int16_t *col, int16_t *line, const uint16_t color, const uint16_t bg, const uint8_t scale);
+void driver_print(const unsigned char* text, const uint32_t len, float *col, float *line, const uint16_t color, const uint16_t bg, const uint8_t scale);
 
 void driver_send_buffer(buffer_data_t buffer_data);
 void driver_buffer_task(void *pvParameters);
